CafeAPI/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==41.0.3
Django==4.2.5
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
jwt==1.3.1
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing-extensions==4.7.1
